I don't want to revert this change to hints.cc
I don't want to revert the changes to hints.cc, just incase the hints depends on the lastest species used in hints mode.
I don't know if that's the case,  but It'd be weird if they were telling you something for a minotaur hunter/ranger and you were playing a halfling ranger.